requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
google==3.0.0
idna==3.7
joblib==1.4.2
numpy==1.26.4
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
soupsieve==2.5
sqlparse==0.5.0
threadpoolctl==3.5.0
typing_extensions==4.11.0
urllib3==2.2.1
